Resolved dependency cycle between actions and web utils
