+++ client/app/scripts/actions/app-actions.js
@@ -1,43 +1,17 @@
-import debug from 'debug';
-import { fromJS } from 'immutable';
-
 import ActionTypes from '../constants/action-types';
 import { saveGraph } from '../utils/file-utils';
 import { clearStoredViewState, updateRoute } from '../utils/router-utils';
-import {
-  doRequest,
-  getApiPath,
-  getAllNodes,
-  getResourceViewNodesSnapshot,
-  getNodeDetails,
-  getTopologies,
-  deletePipe,
-  stopPolling,
-  teardownWebsockets,
-  getNodes,
-} from '../utils/web-api-utils';
 import { isPausedSelector } from '../selectors/time-travel';
 import {
-  availableMetricTypesSelector,
   nextPinnedMetricTypeSelector,
   previousPinnedMetricTypeSelector,
-  pinnedMetricSelector,
 } from '../selectors/node-metric';
-import {
-  isResourceViewModeSelector,
-  resourceViewAvailableSelector,
-} from '../selectors/topology';
+import { isResourceViewModeSelector } from '../selectors/topology';
 
 import {
   GRAPH_VIEW_MODE,
   TABLE_VIEW_MODE,
-  RESOURCE_VIEW_MODE,
 } from '../constants/naming';
-
-
-let controlErrorTimer = 0;
-
-const log = debug('scope:app-actions');
 
 
 export function showHelp() {
@@ -184,39 +158,8 @@ export function updateSearch(searchQuery = '', pinnedSearches = []) {
   };
 }
 
-export function focusSearch() {
-  return (dispatch, getState) => {
-    dispatch({ type: ActionTypes.FOCUS_SEARCH });
-    // update nodes cache to allow search across all topologies,
-    // wait a second until animation is over
-    // NOTE: This will cause matching recalculation (and rerendering)
-    // of all the nodes in the topology, instead applying it only on
-    // the nodes delta. The solution would be to implement deeper
-    // search selectors with per-node caching instead of per-topology.
-    setTimeout(() => {
-      getAllNodes(getState(), dispatch);
-    }, 1200);
-  };
-}
-
 export function blurSearch() {
   return { type: ActionTypes.BLUR_SEARCH };
-}
-
-export function changeTopologyOption(option, value, topologyId, addOrRemove) {
-  return (dispatch, getState) => {
-    dispatch({
-      addOrRemove,
-      option,
-      topologyId,
-      type: ActionTypes.CHANGE_TOPOLOGY_OPTION,
-      value
-    });
-    updateRoute(getState);
-    // update all request workers with new options
-    getTopologies(getState, dispatch);
-    getNodes(getState, dispatch);
-  };
 }
 
 export function clickBackground() {
@@ -224,18 +167,6 @@ export function clickBackground() {
     dispatch({
       type: ActionTypes.CLICK_BACKGROUND
     });
-    updateRoute(getState);
-  };
-}
-
-export function clickCloseDetails(nodeId) {
-  return (dispatch, getState) => {
-    dispatch({
-      nodeId,
-      type: ActionTypes.CLICK_CLOSE_DETAILS
-    });
-    // Pull the most recent details for the next details panel that comes into focus.
-    getNodeDetails(getState, dispatch);
     updateRoute(getState);
   };
 }
@@ -297,102 +228,6 @@ export function setTableView() {
       viewMode: TABLE_VIEW_MODE,
     });
     updateRoute(getState);
-  };
-}
-
-export function setResourceView() {
-  return (dispatch, getState) => {
-    if (resourceViewAvailableSelector(getState())) {
-      dispatch({
-        type: ActionTypes.SET_VIEW_MODE,
-        viewMode: RESOURCE_VIEW_MODE,
-      });
-      // Pin the first metric if none of the visible ones is pinned.
-      const state = getState();
-      if (!pinnedMetricSelector(state)) {
-        const firstAvailableMetricType = availableMetricTypesSelector(state).first();
-        dispatch(pinMetric(firstAvailableMetricType));
-      }
-      getResourceViewNodesSnapshot(getState(), dispatch);
-      updateRoute(getState);
-    }
-  };
-}
-
-export function clickNode(nodeId, label, origin, topologyId = null) {
-  return (dispatch, getState) => {
-    dispatch({
-      label,
-      nodeId,
-      origin,
-      topologyId,
-      type: ActionTypes.CLICK_NODE,
-    });
-    updateRoute(getState);
-    getNodeDetails(getState, dispatch);
-  };
-}
-
-export function pauseTimeAtNow() {
-  return (dispatch, getState) => {
-    dispatch({
-      type: ActionTypes.PAUSE_TIME_AT_NOW
-    });
-    updateRoute(getState);
-    if (!getState().get('nodesLoaded')) {
-      getNodes(getState, dispatch);
-      if (isResourceViewModeSelector(getState())) {
-        getResourceViewNodesSnapshot(getState(), dispatch);
-      }
-    }
-  };
-}
-
-export function clickRelative(nodeId, topologyId, label, origin) {
-  return (dispatch, getState) => {
-    dispatch({
-      label,
-      nodeId,
-      origin,
-      topologyId,
-      type: ActionTypes.CLICK_RELATIVE
-    });
-    updateRoute(getState);
-    getNodeDetails(getState, dispatch);
-  };
-}
-
-function updateTopology(dispatch, getState) {
-  const state = getState();
-  // If we're in the resource view, get the snapshot of all the relevant node topologies.
-  if (isResourceViewModeSelector(state)) {
-    getResourceViewNodesSnapshot(state, dispatch);
-  }
-  updateRoute(getState);
-  // NOTE: This is currently not needed for our static resource
-  // view, but we'll need it here later and it's simpler to just
-  // keep it than to redo the nodes delta updating logic.
-  getNodes(getState, dispatch);
-}
-
-export function clickShowTopologyForNode(topologyId, nodeId) {
-  return (dispatch, getState) => {
-    dispatch({
-      nodeId,
-      topologyId,
-      type: ActionTypes.CLICK_SHOW_TOPOLOGY_FOR_NODE
-    });
-    updateTopology(dispatch, getState);
-  };
-}
-
-export function clickTopology(topologyId) {
-  return (dispatch, getState) => {
-    dispatch({
-      topologyId,
-      type: ActionTypes.CLICK_TOPOLOGY
-    });
-    updateTopology(dispatch, getState);
   };
 }
 
@@ -520,47 +355,10 @@ export function receiveNodesDelta(delta) {
   };
 }
 
-export function resumeTime() {
-  return (dispatch, getState) => {
-    if (isPausedSelector(getState())) {
-      dispatch({
-        type: ActionTypes.RESUME_TIME
-      });
-      updateRoute(getState);
-      // After unpausing, all of the following calls will re-activate polling.
-      getTopologies(getState, dispatch);
-      getNodes(getState, dispatch, true);
-      if (isResourceViewModeSelector(getState())) {
-        getResourceViewNodesSnapshot(getState(), dispatch);
-      }
-    }
-  };
-}
-
 export function receiveNodes(nodes) {
   return {
     nodes,
     type: ActionTypes.RECEIVE_NODES,
-  };
-}
-
-export function jumpToTime(timestamp) {
-  return (dispatch, getState) => {
-    dispatch({
-      timestamp,
-      type: ActionTypes.JUMP_TO_TIME,
-    });
-    updateRoute(getState);
-    getTopologies(getState, dispatch);
-    if (!getState().get('nodesLoaded')) {
-      getNodes(getState, dispatch);
-      if (isResourceViewModeSelector(getState())) {
-        getResourceViewNodesSnapshot(getState(), dispatch);
-      }
-    } else {
-      // Get most recent details before freezing the state.
-      getNodeDetails(getState, dispatch);
-    }
   };
 }
 
@@ -569,53 +367,6 @@ export function receiveNodesForTopology(nodes, topologyId) {
     nodes,
     topologyId,
     type: ActionTypes.RECEIVE_NODES_FOR_TOPOLOGY
-  };
-}
-
-export function receiveTopologies(topologies) {
-  return (dispatch, getState) => {
-    const firstLoad = !getState().get('topologiesLoaded');
-    dispatch({
-      topologies,
-      type: ActionTypes.RECEIVE_TOPOLOGIES
-    });
-    getNodes(getState, dispatch);
-    // Populate search matches on first load
-    const state = getState();
-    // Fetch all the relevant nodes once on first load
-    if (firstLoad && isResourceViewModeSelector(state)) {
-      getResourceViewNodesSnapshot(state, dispatch);
-    }
-  };
-}
-
-export function receiveApiDetails(apiDetails) {
-  return (dispatch, getState) => {
-    const isFirstTime = !getState().get('version');
-    const pausedAt = getState().get('pausedAt');
-
-    dispatch({
-      capabilities: fromJS(apiDetails.capabilities || {}),
-      hostname: apiDetails.hostname,
-      newVersion: apiDetails.newVersion,
-      plugins: apiDetails.plugins,
-      type: ActionTypes.RECEIVE_API_DETAILS,
-      version: apiDetails.version,
-    });
-
-    // On initial load either start time travelling at the pausedAt timestamp
-    // (if it was given as URL param) if time travelling is enabled, otherwise
-    // simply pause at the present time which is arguably the next best thing
-    // we could do.
-    // NOTE: We can't make this decision before API details are received because
-    // we have no prior info on whether time travel would be available.
-    if (isFirstTime && pausedAt) {
-      if (apiDetails.capabilities && apiDetails.capabilities.historic_reports) {
-        dispatch(jumpToTime(pausedAt));
-      } else {
-        dispatch(pauseTimeAtNow());
-      }
-    }
   };
 }
 
@@ -639,86 +390,11 @@ export function receiveControlPipeFromParams(pipeId, rawTty, resizeTtyControl) {
   };
 }
 
-export function receiveControlPipe(pipeId, nodeId, rawTty, resizeTtyControl, control) {
-  return (dispatch, getState) => {
-    const state = getState();
-    if (state.get('nodeDetails').last()
-      && nodeId !== state.get('nodeDetails').last().id) {
-      log('Node was deselected before we could set up control!');
-      deletePipe(pipeId, dispatch);
-      return;
-    }
-
-    const controlPipe = state.get('controlPipes').last();
-    if (controlPipe && controlPipe.get('id') !== pipeId) {
-      deletePipe(controlPipe.get('id'), dispatch);
-    }
-
-    dispatch({
-      control,
-      nodeId,
-      pipeId,
-      rawTty,
-      resizeTtyControl,
-      type: ActionTypes.RECEIVE_CONTROL_PIPE
-    });
-
-    updateRoute(getState);
-  };
-}
-
 export function receiveControlPipeStatus(pipeId, status) {
   return {
     pipeId,
     status,
     type: ActionTypes.RECEIVE_CONTROL_PIPE_STATUS
-  };
-}
-
-function doControlRequest(nodeId, control, dispatch) {
-  clearTimeout(controlErrorTimer);
-  const url = `${getApiPath()}/api/control/${encodeURIComponent(control.probeId)}/`
-    + `${encodeURIComponent(control.nodeId)}/${control.id}`;
-  doRequest({
-    error: (err) => {
-      dispatch(receiveControlError(nodeId, err.response));
-      controlErrorTimer = setTimeout(() => {
-        dispatch(clearControlError(nodeId));
-      }, 10000);
-    },
-    method: 'POST',
-    success: (res) => {
-      dispatch(receiveControlSuccess(nodeId));
-      if (res) {
-        if (res.pipe) {
-          dispatch(blurSearch());
-          const resizeTtyControl = res.resize_tty_control
-            && { id: res.resize_tty_control, nodeId: control.nodeId, probeId: control.probeId };
-          dispatch(receiveControlPipe(
-            res.pipe,
-            nodeId,
-            res.raw_tty,
-            resizeTtyControl,
-            control
-          ));
-        }
-        if (res.removedNode) {
-          dispatch(receiveControlNodeRemoved(nodeId));
-        }
-      }
-    },
-    url
-  });
-}
-
-export function doControl(nodeId, control) {
-  return (dispatch) => {
-    dispatch({
-      control,
-      nodeId,
-      type: ActionTypes.DO_CONTROL
-    });
-    doControlRequest(nodeId, control, dispatch);
   };
 }
 
@@ -747,41 +423,6 @@ export function setContrastMode(enabled) {
   };
 }
 
-export function getTopologiesWithInitialPoll() {
-  return (dispatch, getState) => {
-    getTopologies(getState, dispatch, true);
-  };
-}
-
-export function route(urlState) {
-  return (dispatch, getState) => {
-    dispatch({
-      state: urlState,
-      type: ActionTypes.ROUTE_TOPOLOGY
-    });
-    // Handle Time Travel state update through separate actions as it's more complex.
-    // This is mostly to handle switching contexts Explore <-> Monitor in WC while
-    // the timestamp keeps changing - e.g. if we were Time Travelling in Scope and
-    // then went live in Monitor, switching back to Explore should properly close
-    // the Time Travel etc, not just update the pausedAt state directly.
-    if (!urlState.pausedAt) {
-      dispatch(resumeTime());
-    } else {
-      dispatch(jumpToTime(urlState.pausedAt));
-    }
-    // update all request workers with new options
-    getTopologies(getState, dispatch);
-    getNodes(getState, dispatch);
-    // If we are landing on the resource view page, we need to fetch not only all the
-    // nodes for the current topology, but also the nodes of all the topologies that make
-    // the layers in the resource view.
-    const state = getState();
-    if (isResourceViewModeSelector(state)) {
-      getResourceViewNodesSnapshot(state, dispatch);
-    }
-  };
-}
-
 export function resetLocalViewState() {
   return (dispatch) => {
     dispatch({ type: ActionTypes.RESET_LOCAL_VIEW_STATE });
@@ -804,16 +445,6 @@ export function changeInstance() {
       type: ActionTypes.CHANGE_INSTANCE
     });
     updateRoute(getState);
-  };
-}
-
-export function shutdown() {
-  return (dispatch) => {
-    stopPolling();
-    teardownWebsockets();
-    dispatch({
-      type: ActionTypes.SHUTDOWN
-    });
   };
 }
