+++ client/app/scripts/components/terminal.js
@@ -8,10 +8,11 @@ import { Terminal as Term } from 'xterm';
 import * as fit from 'xterm/lib/addons/fit/fit';
 
 import { closeTerminal } from '../actions/app-actions';
+import { getPipeStatus } from '../actions/request-actions';
 import { getNeutralColor } from '../utils/color-utils';
 import { setDocumentTitle } from '../utils/title-utils';
 import {
-  getPipeStatus, deletePipe, doResizeTty, getWebsocketUrl, basePath
+  deletePipe, doResizeTty, getWebsocketUrl, basePath
 } from '../utils/web-api-utils';
 
 const log = debug('scope:terminal');
