+++ client/app/scripts/components/topologies.js
@@ -5,7 +5,7 @@ import classnames from 'classnames';
 import { trackAnalyticsEvent } from '../utils/tracking-utils';
 import { searchMatchCountByTopologySelector } from '../selectors/search';
 import { isResourceViewModeSelector } from '../selectors/topology';
-import { clickTopology } from '../actions/app-actions';
+import { clickTopology } from '../actions/request-actions';
 
 
 function basicTopologyInfo(topology, searchMatchCount) {
